feat: wip protocol rewrite
